multimodal-matching/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# FAISS (choose one based on hardware)
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # uncomment if GPU available

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data & utils
numpy>=1.24.0
Pillow>=9.0.0
PyYAML>=6.0
requests>=2.28.0
tqdm>=4.65.0

# Keyword extraction
keybert>=0.8.0

# Lexical retrieval (BM25 hybrid)
rank-bm25>=0.2.2

# Dev / notebook
jupyter>=1.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
